refactor: SurveyList의 스타일을 분리합니다.
